address to be delivered is stored

+++ dealmart/buyer/views.py
@@ -76,33 +76,36 @@ class Activate(APIView):
 
 
     def post(self,request,user_id,*args,**kwargs):
-        code = request.data.get('otp')
-        print(user_id)
-        try:
-            otp = OTP.objects.get(receiver=user_id)
-        except(TypeError, ValueError, OverflowError, otp.DoesNotExist):
-            otp = None
-        try:
-            receiver = User.objects.get(id=user_id)
-        except(TypeError, ValueError, OverflowError, User.DoesNotExist):
-            receiver = None
-        if otp is None or receiver is None:
-            return Response({'error':'you are not a valid user'},status=status.HTTP_400_BAD_REQUEST)
+        code = OTPSerializer(data=request.data)
+        if code.is_valid(raise_exception=True):
+            print(user_id)
+            try:
+                otp = OTP.objects.get(receiver=user_id)
+            except(TypeError, ValueError, OverflowError, otp.DoesNotExist):
+                otp = None
+            try:
+                receiver = User.objects.get(id=user_id)
+            except(TypeError, ValueError, OverflowError, User.DoesNotExist):
+                receiver = None
+            if otp is None or receiver is None:
+                return Response({'error':'you are not a valid user'},status=status.HTTP_400_BAD_REQUEST)
 
-        elif timezone.now() - otp.sent_on >= timedelta(days=0,hours=0,minutes=2,seconds=0):
-            otp.delete()
-            return Response({'detail':'OTP expired!',
-                             'user_id':user_id})
+            elif timezone.now() - otp.sent_on >= timedelta(days=0,hours=0,minutes=2,seconds=0):
+                otp.delete()
+                return Response({'detail':'OTP expired!',
+                                 'user_id':user_id})
 
-        if otp.otp == int(code):
-            receiver.is_active = True
-            receiver.save()
-            login(request, receiver)
-            otp.delete()
-            return Response({'message': 'Thank you for Email Verification you are successfully logged in'},
+            if otp.otp == int(code):
+                receiver.is_active = True
+                receiver.save()
+                login(request, receiver)
+                otp.delete()
+                return Response({'message': 'Thank you for Email Verification you are successfully logged in'},
                             status=status.HTTP_200_OK)
-        else:
-            return Response({'error':'Invalid OTP',},status = status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
+            else:
+                return Response({'error':'Invalid OTP',},status = status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
+        return Response({'error':'Invalid OTP',},status = status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
+
 
 
 class ResendOtp(generics.CreateAPIView):
@@ -141,46 +144,51 @@ class Login(APIView):
     permission_classes = (permissions.AllowAny,)
 
     def post(self,request,*args,**kwargs):
-        uname_or_em = request.data.get('uname_or_em')
-        if not uname_or_em:
-            return Response({'detail':'Email or Username is required'},
-                            status=status.HTTP_406_NOT_ACCEPTABLE)
-        password = request.data.get('password')
-        if not password:
-            return Response({'error':'please enter the password!'},
-                            status=status.HTTP_406_NOT_ACCEPTABLE)
-        user = EmailOrUsername(self,uname_or_em = uname_or_em,password=password)
+        serializer = LoginSerializer(data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            uname_or_em = serializer.validated_data['uname_or_em']
+            password = serializer.validated_data['password']
+            user = EmailOrUsername(self,uname_or_em = uname_or_em,password=password)
 
-        if user == 2:
-            return Response({'error':'Invalid Username or Email!!'},
-                            status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
-        elif user == 3:
-            return Response({'error':'Incorrect Password'},
-                            status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
-        else:
-            if user.is_active:
-                login(request, user)
-                return Response({'detail':'successfully Logged in!','user_id': user.id})
+            if user == 2:
+                return Response({'error':'Invalid Username or Email!!'},
+                                status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
+            elif user == 3:
+                return Response({'error':'Incorrect Password'},
+                                status=status.HTTP_203_NON_AUTHORITATIVE_INFORMATION)
             else:
-                return Response({'error':'Please! varify Your Email First','user_id':user.id},
-                                status=status.HTTP_406_NOT_ACCEPTABLE)
+                if user.is_active:
+                   login(request, user)
+                   return Response({'detail':'successfully Logged in!','user_id': user.id})
+                else:
+                   return Response({'error':'Please! varify Your Email First','user_id':user.id},
+                                    status=status.HTTP_406_NOT_ACCEPTABLE)
 
 
 class Logout(APIView):
     """
-    logout view.Only authenticated user can access this url
+    logout view.Only authenticated user can access this url(by default)
 
     """
-    permission_classes = (permissions.IsAuthenticated,)
-
     def get(self,request,*args,**kwargs):
         logout(request)
         return Response({'message':'successfully logged out'},
                         status=status.HTTP_200_OK)
 
 
-class Address(APIView):
+class AddressView(generics.ListCreateAPIView):
     serializer_class = AddressSerializer
+    queryset = Address.objects.all()
+
+    def post(self,request,*args,**kwargs):
+        address = AddressSerializer(data=request.data)
+        if address.is_valid(raise_exception=True):
+            address.save(user=request.user)
+            return Response({'message':'address saved successfully'},
+                              status=status.HTTP_200_OK)
+        return Response({'error':'Not a valid address!'},
+                         status = status.HTTP_406_NOT_ACCEPTABLE)
+
